requirements.txt
fastapi
uvicorn
pydantic
openai
azure-cosmos
azure-storage-blob
azure-ai-formrecognizer
azure-core
streamlit
PyPDF2
python-dotenv
requests
pandas
python-dateutil
typing-extensions
tiktoken
streamlit-lottie
aiohttp